requirements.txt
flask==3.1.0
anthropic==0.94.0
yfinance==0.2.54
pdfplumber==0.11.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
